# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Directive OS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /agencies/me:
    get:
      operationId: getMyAgency
      tags: [agencies]
      summary: Get current agency
      responses:
        "200":
          description: Agency details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agency"
        "404":
          description: Agency not found
    patch:
      operationId: updateMyAgency
      tags: [agencies]
      summary: Update agency details
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAgencyBody"
      responses:
        "200":
          description: Updated agency
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agency"
  /agencies/onboard:
    post:
      operationId: onboardAgency
      tags: [agencies]
      summary: Onboard a new agency
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OnboardAgencyBody"
      responses:
        "201":
          description: Agency created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agency"
...
  /billing/checkout/setup:
    post:
      operationId: createSetupCheckout
      tags: [billing]
      summary: Create Stripe checkout for $1500 setup fee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutBody"
      responses:
        "200":
          description: Checkout URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutResponse"
  /billing/checkout/subscription:
    post:
      operationId: createSubscriptionCheckout
      tags: [billing]
      summary: Create Stripe subscription checkout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriptionCheckoutBody"
      responses:
        "200":
          description: Checkout URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutResponse"
...
  /ai/chat:
    post:
      operationId: aiChat
      tags: [ai]
      summary: AI receptionist chat message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiChatBody"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiChatResponse"
  /ai/send-form:
    post:
      operationId: aiSendForm
      tags: [ai]
      summary: Email tenancy application form to prospect
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendFormBody"
      responses:
        "200":
          description: Form sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendFormResponse"
...
components:
  schemas:
...
    Agency:
      type: object
      properties:
        id:
          type: integer
        clerkOrgId:
          type: string
        name:
          type: string
        abn:
          type: string
        contactEmail:
          type: string
        contactPhone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        stripeCustomerId:
          type: ["string", "null"]
        subscriptionStatus:
          type: string
          enum: [pending, active, past_due, canceled, trialing]
        setupFeePaid:
          type: boolean
        termsAccepted:
          type: boolean
        seatCount:
          type: integer
        aiMinutesUsed:
          type: integer
        aiMinutesIncluded:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - clerkOrgId
        - name
        - abn
        - contactEmail
        - subscriptionStatus
        - setupFeePaid
        - termsAccepted
        - seatCount
        - aiMinutesUsed
        - aiMinutesIncluded
        - createdAt
    UpdateAgencyBody:
      type: object
      properties:
        name:
          type: string
        contactEmail:
          type: string
        contactPhone:
          type: string
        address:
          type: string
        logoUrl:
          type: string
    OnboardAgencyBody:
      type: object
      properties:
        clerkOrgId:
          type: string
        name:
          type: string
        abn:
          type: string
        contactEmail:
          type: string
        contactPhone:
          type: string
        seatCount:
          type: integer
        termsAccepted:
          type: boolean
      required:
        - clerkOrgId
        - name
        - abn
        - contactEmail
        - seatCount
        - termsAccepted
...
    CheckoutBody:
      type: object
      properties:
        successUrl:
          type: string
        cancelUrl:
          type: string
      required:
        - successUrl
        - cancelUrl
    SubscriptionCheckoutBody:
      type: object
      properties:
        seatCount:
          type: integer
        successUrl:
          type: string
        cancelUrl:
          type: string
      required:
        - seatCount
        - successUrl
        - cancelUrl
    CheckoutResponse:
      type: object
      properties:
        url:
          type: string
      required:
        - url
...
    AiChatBody:
      type: object
      properties:
        sessionId:
          type: string
        message:
          type: string
        agencyId:
          type: ["integer", "null"]
      required:
        - sessionId
        - message
    AiChatResponse:
      type: object
      properties:
        reply:
          type: string
        action:
          type: ["string", "null"]
          enum: [send_form, transfer_call, collect_lead, null]
        leadData:
          type: object
          properties:
            name:
              type: string
            email:
              type: string
            phone:
              type: string
            leadType:
              type: string
          additionalProperties: true
      required:
        - reply
    SendFormBody:
      type: object
      properties:
        email:
          type: string
        name:
          type: string
        listingAddress:
          type: string
        agencyName:
          type: string
      required:
        - email
        - name
    SendFormResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message
...