# lib/api-zod/src/generated/types/sendFormResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Directive OS API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SendFormResponse {
  success: boolean;
  message: string;
}